Python project
This project contains the documentation of an in-depth analysis of a police killing dataset found on Kaggle using Python. The dataset provides information on victims of police killings between the months; of January to June 2015 in America.

The goal of this project was to suggest factors that led to increased rates of police killings and generate smart insights to reduce future occurrences.